Moved the extra shade stuff into non-extra module
