+++ src/render/src/shade.rs
@@ -14,7 +14,8 @@
 
 //! Shader parameter handling.
 
-pub use gfx_core::shade::{ConstFormat, Formatted, UniformValue};
+pub use gfx_core::shade::{ConstFormat, Formatted};
+use gfx_core::shade::UniformValue;
 
 pub trait ToUniform: Copy {
     fn convert(self) -> UniformValue;
@@ -44,3 +45,59 @@ uniform!([f32; 4], F32Vector4);
 uniform!([[f32; 2]; 2], F32Matrix2);
 uniform!([[f32; 3]; 3], F32Matrix3);
 uniform!([[f32; 4]; 4], F32Matrix4);
+
+
+use gfx_core::shade::{CreateShaderError, CreateProgramError, ShaderModel};
+
+/// Program linking error
+#[derive(Clone, PartialEq, Debug)]
+pub enum ProgramError {
+    /// Unable to compile the vertex shader
+    Vertex(CreateShaderError),
+    /// Unable to compile the pixel shader
+    Pixel(CreateShaderError),
+    /// Unable to link
+    Link(CreateProgramError),
+}
+
+/// A type storing shader source for different graphics APIs and versions.
+#[allow(missing_docs)]
+#[derive(Copy, Clone, PartialEq, Debug)]
+pub struct ShaderSource<'a> {
+    pub glsl_120: Option<&'a [u8]>,
+    pub glsl_130: Option<&'a [u8]>,
+    pub glsl_140: Option<&'a [u8]>,
+    pub glsl_150: Option<&'a [u8]>,
+    pub glsl_430: Option<&'a [u8]>,
+    // TODO: hlsl_sm_N...
+    pub targets: &'a [&'a str],
+}
+
+impl<'a> ShaderSource<'a> {
+    /// Create an empty shader source. Useful for specifying the remaining
+    /// structure members upon construction.
+    pub fn empty() -> ShaderSource<'a> {
+        ShaderSource {
+            glsl_120: None,
+            glsl_130: None,
+            glsl_140: None,
+            glsl_150: None,
+            glsl_430: None,
+            targets: &[],
+        }
+    }
+
+    /// Pick one of the stored versions that is the highest supported by the device.
+    pub fn choose(&self, model: ShaderModel) -> Result<&'a [u8], ()> {
+        // following https://www.opengl.org/wiki/Detecting_the_Shader_Model
+        let version = model.to_number();
+        Ok(match *self {
+            ShaderSource { glsl_430: Some(s), .. } if version >= 50 => s,
+            ShaderSource { glsl_150: Some(s), .. } if version >= 40 => s,
+            ShaderSource { glsl_140: Some(s), .. } if version >= 40 => s,
+            ShaderSource { glsl_130: Some(s), .. } if version >= 30 => s,
+            ShaderSource { glsl_120: Some(s), .. } if version >= 20 => s,
+            _ => return Err(()),
+        })
+    }
+}
